requirements.txt
pandas==2.2.2
pillow==10.3.0
tensorflow==2.16.1
streamlit==0.79.0
numpy==1.19.5